moved error logic to correct place (old logic still in main)
