+++ octizys_parser/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod lexer;
 pub mod parser;
 pub mod report;
+pub mod tokens;
 use lalrpop_util::lalrpop_mod;
 lalrpop_mod!(pub grammar);
